### 2024-03-30 #6555 @danielskatz
@victorapm - thanks for your submission to JOSS.  Your paper is quite short, and it might benefit from adding a few publications in which this work was used, and/or some examples of how it might be used, and for what types of problems.  Also, and as a minor point, "a software" should be replaced by "a software library" or "a software package". Software is an uncountable noun.
