### 2024-03-30 #6555 @danielskatz
Please feel free to make changes to your .bib and/or .md files, then use the command `@editorialbot check references` to check the references, and/or the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
