+++ docs/joss/paper.bib
@@ -4,3 +4,34 @@
   url    = {https://github.com/hypre-space/hypre},
   year   = {2024},
 }
+
+@misc{YAML,
+  author = {{YAML}},
+  title  = {{YAML Ain't Markup Language (YAML) Version 1.2}},
+  url    = {https://yaml.org/spec/1.2/spec.html},
+  year   = {2009},
+}
+
+@inproceedings{BrKnMaMcSm12,
+  author    = {Brown, Jed and Knepley, Matthew G. and May, David A. and McInnes, Lois Curfman and Smith, Barry},
+  booktitle = {2012 11th International Symposium on Parallel and Distributed Computing},
+  title     = {Composable Linear Solvers for Multiphysics},
+  year      = {2012},
+  pages     = {55-62},
+  keywords  = {Jacobian matrices;Physics;Indexes;Ice;Sparse matrices;Runtime;Couplings;multiphysics;hierarchical solvers;composable solvers;Schur complement},
+  doi       = {10.1109/ISPDC.2012.16}
+}
+
+@inproceedings{spack,
+  address = {Austin, Texas, USA},
+  author  = {Gamblin, Todd and LeGendre, Matthew and
+             Collette, Michael R. and Lee, Gregory L. and
+             Moody, Adam and de Supinski, Bronis R. and Futral, Scott},
+  doi     = {10.1145/2807591.2807623},
+  month   = {November 15-20},
+  note    = {LLNL-CONF-669890},
+  series  = {Supercomputing 2015 (SC’15)},
+  title   = {{The Spack Package Manager: Bringing Order to HPC Software Chaos}},
+  url     = {https://github.com/spack/spack},
+  year    = {2015}
+}
