+++ docs/joss/paper.md
@@ -12,47 +12,114 @@ authors:
 affiliations:
  - name: Lawrence Livermore National Laboratory, CA, USA
    index: 1
-date: 1 April 2024
+date: 30 March 2024
 bibliography: paper.bib
 ---
 
 # Summary
 
-Solving sparse linear systems of equations is an essential problem to many application
+Solving sparse linear systems of equations is an essential problem for many application
 codes in computational science and engineering (CSE). *Hypredrive* aims to facilitate this
 problem via a simple and user-friendly interface to *hypre* [@hypre], a well-established
-package featuring multigrid methods. *Hypredrive* allows users to easily configure and
-switch solver options in *hypre* through YAML input files, making experimentation with
-different solver techniques more accessible to researchers and software developers working
-with numerical simulation codes.
+package featuring multigrid methods. Inspired by the solver composability work done in
+PETSc [@BrKnMaMcSm12] and flexible configuration framework used in [@spack], *hypredrive*
+allows users to easily configure and switch solver options in *hypre* through [@YAML]
+input files, making experimentation with different solver techniques more accessible to
+researchers and software developers who work with numerical simulation codes.
 
 # Statement of need
 
 *Hypre* is a widely used and efficient linear solver package; however, the complexity
-associated with its direct use may make it difficult to experiment with different solver
+associated with its direct use might limit the exploration with different solver
 options. *Hypredrive* bridges this gap by providing a high-level and lightweight interface
 to *hypre*, encapsulating its complexity while retaining its capabilities with minimal
 computational overhead.
 
 # Software capabilities
 
-*Hypredrive* is a software written in C that includes a library with APIs designed to
-simplify the interaction with *hypre* and an executable for performing the solution of
+*Hypredrive* is a software package written in C that includes a library with APIs designed
+to simplify the interaction with *hypre* and an executable for performing the solution of
 linear systems defined via YAML input files. Key features of the software are:
 
 * **Encapsulation**: `libHYPREDRV` wraps the function calls for building solvers and
   preconditioners in *hypre* through an intuitive YAML interface driven by configuration
-  files.
+  files. This design ensures a straightforward way of setting up solvers in *hypre* and
+  sharing options without recompiling the user's application code.
 
 
 * **Prototyping**: `hypredrive` allows users to prototype rapidly, comparing the
-  performance of various solver options updated directly through the YAML configuration file.
+  performance of various solver options and tweaking parameters directly through the YAML
+  configuration file. This flexibility encourages experimentation, helping users identify
+  the most effective solver strategies for their specific problems.
 
 
 * **Testing**: `hypredrive` enables the creation of an integrated testing framework to
-  evaluate solvers against a set of predefined linear systems.
+  evaluate solvers against a set of predefined linear systems. This feature lets users
+  understand whether updates to *hypre* lead to different solver convergence or
+  performance for their problems of interest.
 
-For instructions on installation, input file structure, and examples, see [*hypredrive*'s
+# Example usage
+
+As an example usage of `hypredrive`, consider the solution of a linear system arising from
+a seven points finite difference discretization of the Laplace equation on a `10x10x10`
+cartesian grid. Both sparse matrix and right hand side vector are stored at
+`data/ps3d10pt7/np1/`. For solving this linear system with algebraic multigrid (BoomerAMG)
+as a preconditioner to the conjugate gradient iterative solver, the YAML input file
+`input.yml` would look like:
+
+```yaml
+linear_system:
+  rhs_filename: data/ps3d10pt7/np1/IJ.out.b
[…]
