+++ docs/joss/paper.md
@@ -93,26 +93,26 @@ Date and time: YYYY-MM-DD HH:MM:SS
 Using HYPRE_DEVELOP_STRING: HYPRE_VERSION_GOES_HERE
 
 Running on 1 MPI rank
------------------------------------------------------------------------------------
+-------------------------------------------------------------------------------
 linear_system:
   rhs_filename: data/ps3d10pt7/np1/IJ.out.b
   matrix_filename: data/ps3d10pt7/np1/IJ.out.A
 solver: pcg
 preconditioner: amg
------------------------------------------------------------------------------------
-===================================================================================
+-------------------------------------------------------------------------------
+===============================================================================
 Solving linear system #0 with 1000 rows and 6400 nonzeros...
-===================================================================================
+===============================================================================
 
 
 STATISTICS SUMMARY:
 
-+------------+-------------+-------------+-------------+-------------+------------+
-|            |    LS build |       setup |       solve |    relative |            |
-|      Entry |       times |       times |       times |   res. norm |      iters |
-+------------+-------------+-------------+-------------+-------------+------------+
-|          0 |       0.004 |       0.002 |       0.001 |    4.98e-08 |          6 |
-+------------+-------------+-------------+-------------+-------------+------------+
++----------+-------------+-------------+-------------+-------------+----------+
+|          |    LS build |       setup |       solve |    relative |          |
+|    Entry |       times |       times |       times |   res. norm |    iters |
++----------+-------------+-------------+-------------+-------------+----------+
+|        0 |       0.004 |       0.002 |       0.001 |    4.98e-08 |        6 |
++----------+-------------+-------------+-------------+-------------+----------+
 
 Date and time: YYYY-MM-DD HH:MM:SS
 ${HYPREDRIVE_PATH}/hypredrive done!
