### 2024-04-02 4f858da
Update install instructions
