### 2024-04-03 a0fff39
Update install instructions
