### 2024-04-03 97d41ae
Update build framework (#58)
* Remove config.h.in file from version control
* Remove autogen script
* Add with-hypre-dir option
* Update install instructions
* Check for srun or mpirun
* Improve checking for dependency libraries
