### 2024-04-03 c3ad856
Additional changes to build process (#59)
* Revert make check changes from last PR
* configure checks for minimum hypre version
