### 2024-04-09 d116bc8
Add support to HYPRE_MGRSetGlobalSmootherAtLevel
