+++ src/mgr.c
@@ -415,6 +415,7 @@ MGRCreate(MGR_args *args, HYPRE_Solver *precon_ptr, HYPRE_Solver *csolver_ptr)
    HYPRE_Solver   precon;
    HYPRE_Solver   csolver;
    HYPRE_Solver   frelax;
+   HYPRE_Solver   grelax;
    IntArray      *dofmap;
    HYPRE_Int      num_dofs;
    HYPRE_Int      num_dofs_last;
@@ -486,13 +487,23 @@ MGRCreate(MGR_args *args, HYPRE_Solver *precon_ptr, HYPRE_Solver *csolver_ptr)
       HYPRE_MGRSetFSolver(precon, HYPRE_BoomerAMGSolve, HYPRE_BoomerAMGSetup, frelax);
    }
 
-   /* Config f-relaxation at level > 0 (valid only for AMG for now) */
+   /* Config f-relaxation at level > 0 */
    for (i = 1; i < num_levels; i++)
    {
       if (args->level[i].f_relaxation.type == 2)
       {
          AMGCreate(&args->level[i].f_relaxation.amg, &frelax);
          HYPRE_MGRSetFSolverAtLevel(i, precon, frelax);
+      }
+   }
+
+   /* Config global relaxation at level >= 0 */
+   for (i = 0; i < num_levels; i++)
+   {
+      if (args->level[i].g_relaxation.type == 16)
+      {
+         ILUCreate(&args->level[i].g_relaxation.ilu, &grelax);
+         HYPRE_MGRSetGlobalSmootherAtLevel(i, precon, grelax);
       }
    }
 
