### 2024-04-18 #6555 @lucydot
@mayrmt @upsj @DamynChipman - hello 👋  - are you available to review this submission to JOSS?
