### 2024-04-22 #6654 @editorialbot
[REVIEW]: Hypredrive: High-level interface for solving linear systems with hypre
**Submitting author:** @victorapm (<a href="http://orcid.org/0000-0002-3389-523X">Victor Antonio Paludetto Magri</a>)
**Repository:** https://github.com/hypre-space/hypredrive
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.0
