### 2024-04-26 fc30054
Move MGR_LEVEL_ATTRIBUTE macro to mgr.h
