### 2024-04-26 0c110dc
Add option to build RHS given a random solution vector
