### 2024-05-03 hypre-space/hypredrive#60 @DamynChipman
JOSS Paper Comments
In reference to JOSS review thread: https://github.com/openjournals/joss-reviews/issues/6654.

With regard to the state of the field, are there other solver libraries/languages that have quick solver prototyping? For example, MATLAB has rapid prototyping for choosing solvers and options. What specifically about PETSc inspires this work?

With regard to the statement of need, who is the target audience (`hypre` developers, `hypre` users, external users, application developers)?
