+++ docs/joss/paper.bib
@@ -12,14 +12,33 @@
   year   = {2009},
 }
 
-@inproceedings{BrKnMaMcSm12,
-  author    = {Brown, Jed and Knepley, Matthew G. and May, David A. and McInnes, Lois Curfman and Smith, Barry},
-  booktitle = {2012 11th International Symposium on Parallel and Distributed Computing},
-  title     = {Composable Linear Solvers for Multiphysics},
-  year      = {2012},
-  pages     = {55-62},
-  keywords  = {Jacobian matrices;Physics;Indexes;Ice;Sparse matrices;Runtime;Couplings;multiphysics;hierarchical solvers;composable solvers;Schur complement},
-  doi       = {10.1109/ISPDC.2012.16}
+@manual{trilinos,
+  title  = {The {T}rilinos {P}roject {W}ebsite},
+  author = {The {T}rilinos {P}roject {T}eam}},
+  year   = {2024},
+  url    = {https://trilinos.github.io}
+}
+
+@misc{petscrc,
+  author = {{PETSc}},
+  title  = {{PETSc} - Options database},
+  url    = {https://petsc.org/main/manual/other/#the-options-database},
+  year   = {2024},
+}
+
+@misc{petsc,
+  author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed
+           Brown and Peter Brune and Kris Buschelman and Emil~M. Constantinescu and Lisandro
+           Dalcin and Alp Dener and Victor Eijkhout and Jacob Faibussowitsch and William~D.
+           Gropp and V\'{a}clav Hapla and Tobin Isaac and Pierre Jolivet and Dmitry Karpeev
+           and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott Kruger and
+           Dave~A. May and Lois Curfman McInnes and Richard Tran Mills and Lawrence Mitchell
+           and Todd Munson and Jose~E. Roman and Karl Rupp and Patrick Sanan and Jason Sarich
+           and Barry~F. Smith and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao
+           Zhang},
+  title  = {{PETS}c {W}eb page},
+  url    = {https://petsc.org/},
+  year   = {2024}
 }
 
 @inproceedings{spack,
