+++ docs/joss/paper.bib
@@ -13,8 +13,8 @@
 }
 
 @manual{trilinos,
-  title  = {The {T}rilinos {P}roject {W}ebsite},
-  author = {The {T}rilinos {P}roject {T}eam}},
+  title  = {{The Trilinos Project Website}},
+  author = {{The Trilinos Project Team}},
   year   = {2024},
   url    = {https://trilinos.github.io}
 }
