+++ docs/joss/paper.bib
@@ -1,5 +1,5 @@
 @misc{hypre,
-  author = {hypre},
+  author = {{hypre}},
   title  = {high performance preconditioners},
   url    = {https://github.com/hypre-space/hypre},
   year   = {2024},
@@ -12,9 +12,9 @@
   year   = {2009},
 }
 
-@manual{trilinos,
+@misc{trilinos,
+  author = {{Trilinos}},
   title  = {{The Trilinos Project Website}},
-  author = {{The Trilinos Project Team}},
   year   = {2024},
   url    = {https://trilinos.github.io}
 }
