+++ docs/joss/paper.bib
@@ -21,7 +21,7 @@
 
 @misc{petscrc,
   author = {{PETSc}},
-  title  = {{PETSc} - Options database},
+  title  = {The options database},
   url    = {https://petsc.org/main/manual/other/#the-options-database},
   year   = {2024},
 }
