### 2024-05-06 3628bcc
Update JOSS paper (#61)
* Addressing review
* Fix references
* Tweaks
* Rewording
