### 2024-05-07 2697af5
Add HYPRE_CHECK_MIN_VERSION
