### 2024-05-07 d5670e5
Add option to print times in milliseconds
