### 2024-05-27 #6654 @mayrmt
@danielskatz thanks for asking! I have just re-done my tests of the software. One item remains open, that requires some clarification from the submitting author.

@victorapm While the code compiles and tests run nicely, one line of the installation instructions give me some trouble. In particular, `make install` yields the following terminal output:

```bash
$ make install
make[1]: Entering directory '/hdd/codes/hypredrive/src-hyperdrive'
 /usr/bin/mkdir -p '/usr/local/lib'
 /bin/bash ./libtool   --mode=install /usr/bin/install -c   libHYPREDRV.la '/usr/local/lib'
libtool: install: /usr/bin/install -c .libs/libHYPREDRV.lai /usr/local/lib/libHYPREDRV.la
/usr/bin/install: cannot create regular file '/usr/local/lib/libHYPREDRV.la': Permission denied
make[1]: *** [Makefile:532: install-libLTLIBRARIES] Error 1
make[1]: Leaving directory '/hdd/codes/hypredrive/src-hyperdrive'
make: *** [Makefile:1101: install-am] Error 2
```

I was wondering, whether I can change the installation directory, though I haven't found the necessary information in neither the documentation nor the code base itself. Am I missing something here?
