### 2024-05-28 #6654 @victorapm
Hi Matthias,

`make install` failed because, by default, it installs the software at `/usr/local`, which requires admin privileges.

For a custom installation path, you can pass the option `--prefix=${INSTALL_PATH}` to the configure script, where `${INSTALL_PATH}` is a destination folder set by the user, e.g., `INSTALL_PATH=${PWD}/install`. This is mentioned in the "Getting Started" section, fourth enumerated point under "Note".

Hope this helps!
