### 2024-05-30 #6654 @lucydot
Excellent work @mayrmt and @DamynChipman; on behalf of the JOSS team, thank you for your expertise and time.
