### 2024-05-30 #6654 @lucydot
@victorapm you have a very nicely written and concise JOSS paper 👍
One minor point - you do not use the acronym `CSE` defined at the start of the paper.
