### 2024-05-30 #6654 @lucydot
@victorapm a little work is also needed on the references.
I can see all but one reference points to a webpage. Are there outputs-with-a-DOI associated with any of these projects? As these would be our preferred sources. You can still link directly to any webpages directly by embedding a hyperlink in the main text.
