### 2024-06-04 #6654 @lucydot
@victorapm a reminder ping. Few relatively small changes needed to paper and then we are ready to proceed with final steps.
