+++ docs/joss/paper.bib
@@ -1,44 +1,50 @@
-@misc{hypre,
-  author = {{hypre}},
-  title  = {high performance preconditioners},
-  url    = {https://github.com/hypre-space/hypre},
-  year   = {2024},
+@InProceedings{hypre,
+  author="Falgout, Robert D. and Yang, Ulrike Meier",
+  editor="Sloot, Peter M. A. and Hoekstra, Alfons G. and Tan, C. J. Kenneth and Dongarra, Jack J.",
+  title="hypre: A Library of High Performance Preconditioners",
+  booktitle="Computational Science --- ICCS 2002",
+  year="2002",
+  publisher="Springer Berlin Heidelberg",
+  address="Berlin, Heidelberg",
+  pages="632--641",
+  isbn="978-3-540-47789-1",
+  doi="10.1007/3-540-47789-6_66"
 }
 
-@misc{YAML,
-  author = {{YAML}},
-  title  = {{YAML Ain't Markup Language (YAML) Version 1.2}},
-  url    = {https://yaml.org/spec/1.2/spec.html},
-  year   = {2009},
+@article{trilinos,
+  author = {Heroux, Michael A. and Bartlett, Roscoe A. and Howle, Vicki E. and Hoekstra, Robert J. and Hu, Jonathan J. and Kolda, Tamara G. and Lehoucq, Richard B. and Long, Kevin R. and Pawlowski, Roger P. and Phipps, Eric T. and Salinger, Andrew G. and Thornquist, Heidi K. and Tuminaro, Ray S. and Willenbring, James M. and Williams, Alan and Stanley, Kendall S.},
+  title = {An overview of the Trilinos project},
+  year = {2005},
+  issue_date = {September 2005},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  volume = {31},
+  number = {3},
+  issn = {0098-3500},
+  url = {https://doi.org/10.1145/1089014.1089021},
+  doi = {10.1145/1089014.1089021},
+  journal = {ACM Trans. Math. Softw.},
+  month = {sep},
+  pages = {397–423},
+  numpages = {27},
+  keywords = {Software Quality Engineering, Software framework, interfaces}
 }
 
-@misc{trilinos,
-  author = {{Trilinos}},
-  title  = {{The Trilinos Project Website}},
-  year   = {2024},
-  url    = {https://trilinos.github.io}
-}
-
-@misc{petscrc,
-  author = {{PETSc}},
-  title  = {The options database},
-  url    = {https://petsc.org/main/manual/other/#the-options-database},
-  year   = {2024},
-}
-
-@misc{petsc,
-  author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed
-           Brown and Peter Brune and Kris Buschelman and Emil~M. Constantinescu and Lisandro
-           Dalcin and Alp Dener and Victor Eijkhout and Jacob Faibussowitsch and William~D.
-           Gropp and V\'{a}clav Hapla and Tobin Isaac and Pierre Jolivet and Dmitry Karpeev
-           and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott Kruger and
-           Dave~A. May and Lois Curfman McInnes and Richard Tran Mills and Lawrence Mitchell
-           and Todd Munson and Jose~E. Roman and Karl Rupp and Patrick Sanan and Jason Sarich
-           and Barry~F. Smith and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao
-           Zhang},
-  title  = {{PETS}c {W}eb page},
-  url    = {https://petsc.org/},
-  year   = {2024}
+@TechReport{petsc,
[…]
