### 2024-06-14 #6654 @victorapm
Hi Lucy, thank you for your kind comments and I apologize for the delay, the last couple of weeks have been busy to me!

I updated the references as you suggested, i.e., by using the ones with DOI.

The package version is `v0.1.0` and published [here](https://github.com/hypre-space/hypredrive/releases/tag/v0.1.0)

As for the zenodo archive, you can find it [here](https://zenodo.org/records/11657386)

Please, let me know if there's something missing from my end.
Thank you!
