### 2024-06-24 #6654 @danielskatz
👋 @victorapm - the paper looks good, but I found two small issues in the bib to correct, as shown in https://github.com/hypre-space/hypredrive/pull/62. Please merge this, or let me know what you disagree with, then we can continue to acceptance and publication.
