### 2024-06-24 79f7a42
Update paper.bib (#62)
