+++ docs/joss/paper.bib
@@ -13,7 +13,7 @@
 
 @article{trilinos,
   author = {Heroux, Michael A. and Bartlett, Roscoe A. and Howle, Vicki E. and Hoekstra, Robert J. and Hu, Jonathan J. and Kolda, Tamara G. and Lehoucq, Richard B. and Long, Kevin R. and Pawlowski, Roger P. and Phipps, Eric T. and Salinger, Andrew G. and Thornquist, Heidi K. and Tuminaro, Ray S. and Willenbring, James M. and Williams, Alan and Stanley, Kendall S.},
-  title = {An overview of the Trilinos project},
+  title = {An overview of the {T}rilinos project},
   year = {2005},
   issue_date = {September 2005},
   publisher = {Association for Computing Machinery},
@@ -56,7 +56,7 @@
   month   = {November 15-20},
   note    = {LLNL-CONF-669890},
   series  = {Supercomputing 2015 (SC’15)},
-  title   = {{The Spack Package Manager: Bringing Order to HPC Software Chaos}},
+  title   = {The {S}pack Package Manager: Bringing Order to {HPC} Software Chaos},
   url     = {https://github.com/spack/spack},
   year    = {2015}
 }
