### 2024-06-24 #6654 @victorapm
Merged, thanks @danielskatz!
