### 2024-06-24 #6654 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
