### 2024-06-24 #6654 @danielskatz
Congratulations to @victorapm (Victor Antonio Paludetto Magri) on your publication!!

JOSS depends on volunteers and we couldn't be successful without you
